+++ src/test/java/de/alpharogroup/collections/pairs/QuattroTest.java
@@ -66,8 +66,9 @@ public class QuattroTest
 		final Quattro<String, Employee, Customer, Brands> second = Quattro
 			.<String, Employee, Customer, Brands> builder().topLeft("left").topRight(employee)
 			.bottomLeft(customer).bottomRight(ferrari).build();
-		final Quattro<Integer, Employee, Customer, Brands> third = new Quattro<>(1, employee,
-			customer, ferrari);
+		Integer topLeft = 1;
+		final Quattro<Integer, Employee, Customer, Brands> third = new Quattro<>(customer, ferrari,
+			topLeft, employee);
 		final Quattro<Integer, Employee, Customer, Brands> fourth = new Quattro<>();
 		fourth.setTopLeft(1);
 		fourth.setTopRight(employee);