+++ src/test/java/de/alpharogroup/comparators/ChainableComparatorTest.java
@@ -43,14 +43,14 @@ import de.alpharogroup.test.objects.Person;
 public class ChainableComparatorTest
 {
 
-	/** For use of the expected result. */
-	boolean expected;
-
 	/** For use of the result of the comparator. */
 	int actual;
 
 	/** The comparator. */
 	ChainableComparator<Employee> comparator;
+
+	/** For use of the expected result. */
+	boolean expected;
 
 	/**
 	 * Test method for {@link ChainableComparator#compare(Object, Object)}