+++ src/test/java/de/alpharogroup/comparators/pairs/KeyMapPairKeyComparatorTest.java
@@ -44,11 +44,11 @@ import de.alpharogroup.test.objects.Person;
 public class KeyMapPairKeyComparatorTest
 {
 
-	/** For use of the expected result. */
-	boolean expected;
-
 	/** For use of the result of the comparator. */
 	int actual;
+
+	/** For use of the expected result. */
+	boolean expected;
 
 	/**
 	 * Test method for {@link KeyMapPairKeyComparator#compare(Object, Object)}