+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## Change log
 ----------------------
 
+Version 4.31
+-------------
+
+ADDED:
+ 
+- new factory methods with initial capacity for create set objects created
+- new factory methods for primitive array types created
+
 Version 4.30.1
 -------------